fix: rename Trand to Trend
